my scene
so the tentya file is the tentacles and the an4.blend (or fbx) should
contain the animation with them, although the animations haven’t worked
in unity. Unity knows there’s an animation, but won’t play it. Also my
whole scene is really low res and I don’t know why?? Will try to work
on it more tomorrow